completed api of user and passenger

+++ my12306_web/src/main/java/com/dbpp/my12306/service/PassengerService.java
@@ -4,7 +4,6 @@ import com.dbpp.my12306.entity.Passenger;
 import com.dbpp.my12306.mapper.PassengerMapper;
 import com.dbpp.my12306.utils.ResponseSet;
 import com.dbpp.my12306.utils.ResultCode;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.transaction.interceptor.TransactionAspectSupport;
@@ -14,8 +13,11 @@ import java.util.List;
 @Service
 public class PassengerService {
 
-	@Autowired
-	private PassengerMapper passengerMapper;
+	private final PassengerMapper passengerMapper;
+
+	public PassengerService(PassengerMapper passengerMapper) {
+		this.passengerMapper = passengerMapper;
+	}
 
 	/**
 	 * Get the number of passengers
@@ -59,12 +61,11 @@ public class PassengerService {
 	 *
 	 * @return the list of users and user info
 	 */
-	public ResponseSet<List<Passenger>> getAll() {
-		String event = "Get all passengers.";
+	public ResponseSet<List<Passenger>> getAll(boolean seeDisable) {
 		ResponseSet<List<Passenger>> ret = new ResponseSet<>();
 		try {
 			ret.setStatus(ResultCode.SUCCESS);
-			ret.setData(passengerMapper.getAll());
+			ret.setData(passengerMapper.getAll(seeDisable));
 		} catch (Exception e) {
 			ret.setData(null);
 			ret.setStatus(ResultCode.EXCEPTION);
@@ -78,11 +79,29 @@ public class PassengerService {
 	 *
 	 * @return the list of passengers and passenger info
 	 */
-	public ResponseSet<List<Passenger>> getAllOf(int userId) {
+	public ResponseSet<List<Passenger>> getAllOf(int userId, boolean seeDisable) {
 		ResponseSet<List<Passenger>> ret = new ResponseSet<>();
 		try {
 			ret.setStatus(ResultCode.SUCCESS);
-			ret.setData(passengerMapper.getAllOf(userId));
+			ret.setData(passengerMapper.getAllOf(userId, seeDisable));
+		} catch (Exception e) {
+			ret.setData(null);
+			ret.setStatus(ResultCode.EXCEPTION);
+			ret.setDetail(e.getMessage());
+		}
+		return ret;
+	}
+
+	/**
+	 * Get all passengers of one user
+	 *
+	 * @return the list of passengers and passenger info
+	 */
+	public ResponseSet<List<Passenger>> getAllOf(String userName, boolean seeDisable) {
+		ResponseSet<List<Passenger>> ret = new ResponseSet<>();
+		try {
+			ret.setStatus(ResultCode.SUCCESS);
+			ret.setData(passengerMapper.getAllOfByName(userName, seeDisable));
 		} catch (Exception e) {
 			ret.setData(null);
 			ret.setStatus(ResultCode.EXCEPTION);
@@ -98,12 +117,20 @@ public class PassengerService {
 	 * @param id user id
 	 * @return the user object. Null if not exits.
 	 */
-	public ResponseSet<Passenger> getById(int id) {
+	public ResponseSet<Passenger> getById(int id, boolean seeDisable) {
 		ResponseSet<Passenger> ret = new ResponseSet<>();
 		try {
-			ret.setData(passengerMapper.getById(id));
-			ret.setStatus(ret.getData() != null ?
-					ResultCode.SUCCESS : ResultCode.NO_RESULT);
+			var psg = passengerMapper.getById(id);
+			if (psg == null) {
+				ret.setStatus(ResultCode.NO_RESULT);
+			} else if (seeDisable || "Y".equals(psg.getAvailable())) {
+				ret.setData(psg);
+				ret.setStatus(ResultCode.SUCCESS);
+			} else {
+				ret.setDetail("This passenger is disabled and normal user can't access");
+				ret.setStatus(ResultCode.INVALID_AUTH);
+			}
+
 		} catch (Exception e) {
 			ret.setData(null);
 			ret.setStatus(ResultCode.EXCEPTION);
@@ -123,13 +150,23 @@ public class PassengerService {
 		Passenger psg = new Passenger(userId, firstName, lastName, kind, gender, idNo);
 		ResponseSet<Integer> ret = new ResponseSet<>();
 		try {
-			ret.setStatus(ResultCode.SUCCESS);
+			ret.setStatus(ResultCode.CREATED);
 			passengerMapper.add(psg);
-			ret.setData(psg.getPassengerId());
+			psg.setAvailable("Y");
+			if (psg.getKind() == null) {
+				psg.setKind("A");
+			}
+			ret.setData(psg.getUserId());
 		} catch (Exception e) {
 			ret.setDetail(null);
-			ret.setStatus(ResultCode.EXCEPTION);
-			ret.setDetail(e.getMessage());
+			if (e.getCause().getMessage().contains("duplicate key")) {
+				ret.setStatus(ResultCode.SUCCESS_IS_HAVE);
+			} else if (e.getCause().getMessage().contains("row contains")) {
+				ret.setStatus(ResultCode.CONS_ERROR);
+			} else {
+				ret.setStatus(ResultCode.EXCEPTION);
+			}
+			ret.setDetail(e.getCause().getMessage().split("详细：")[1]);
 			TransactionAspectSupport.currentTransactionStatus()
 					.setRollbackOnly();
 		}
@@ -137,24 +174,30 @@ public class PassengerService {
 	}
 
 	/**
-	 * Hide passenger from database
+	 * Set passenger to be disable
 	 *
-	 * @param id   psg id to hide
-	 * @return delete or not
+	 * @return completed or not
 	 */
 	@Transactional
-	public ResponseSet<Integer> hide(int id) {
+	public ResponseSet<Integer> disable(int userId, int psgId) {
 		ResponseSet<Integer> ret = new ResponseSet<>();
 		try {
-			int r = passengerMapper.delete(id);
-			if (r == 1) {
-				ret.setDetail("Hide completed.");
-				ret.setStatus(ResultCode.SUCCESS);
+			Passenger psg = passengerMapper.getById(psgId);
+			if (psg.getUserId() != userId) {
+				ret.setDetail("The user doesn't have the privilege " +
+						"to access this passenger");
+				ret.setStatus(ResultCode.INVALID_AUTH);
 			} else {
-				ret.setDetail("No such passenger.");
-				ret.setStatus(ResultCode.FAIL);
+				int r = passengerMapper.disable(psgId);
+				if (r == 1) {
+					ret.setDetail("Hide completed.");
+					ret.setStatus(ResultCode.SUCCESS);
+				} else {
+					ret.setDetail("No such passenger.");
+					ret.setStatus(ResultCode.FAIL);
+				}
+				ret.setData(r);
 			}
-			ret.setData(r);
 		} catch (Exception e) {
 			ret.setData(null);
 			ret.setStatus(ResultCode.EXCEPTION);
@@ -168,23 +211,34 @@ public class PassengerService {
 	/**
 	 * Delete passenger from database
 	 *
-	 * @param id   psg id to delete
+	 * @param id psg id to delete
 	 * @return delete or not
 	 */
 	@Transactional
-	public ResponseSet<Integer> delete(int id) {
-		String event = "Delete passenger: id=" + id + ".";
+	public ResponseSet<Integer> delete(int id, boolean trueDel) {
 		ResponseSet<Integer> ret = new ResponseSet<>();
 		try {
-			int r = passengerMapper.delete(id);
-			if (r == 1) {
-				ret.setDetail("Delete completed.");
-				ret.setStatus(ResultCode.SUCCESS);
+			if (trueDel) {
+				int r = passengerMapper.delete(id);
+				if (r == 1) {
+					ret.setDetail("Delete completed.");
+					ret.setStatus(ResultCode.SUCCESS);
+				} else {
+					ret.setDetail("No such passenger.");
+					ret.setStatus(ResultCode.FAIL);
+				}
+				ret.setData(r);
 			} else {
-				ret.setDetail("No such passenger.");
-				ret.setStatus(ResultCode.FAIL);
+				int r = passengerMapper.disable(id);
+				if (r == 1) {
+					ret.setDetail("Hide completed.");
+					ret.setStatus(ResultCode.SUCCESS);
+				} else {
+					ret.setDetail("No such passenger.");
+					ret.setStatus(ResultCode.FAIL);
+				}
+				ret.setData(r);
 			}
-			ret.setData(r);
 		} catch (Exception e) {
 			ret.setData(null);
 			ret.setStatus(ResultCode.EXCEPTION);
@@ -194,19 +248,4 @@ public class PassengerService {
 		}
 		return ret;
 	}
-
-//	/**
-//	 * Update the user.
-//	 * If the user not exits, change nothing
-//	 * @param user update this user
-//	 * @return the pk of user
-//	 */
-//	long update(User user);
-//
-//	/**
-//	 * Check exist
-//	 * @param id user id to check
-//	 * @return exist or not
-//	 */
-//	boolean exist(int id);
 }
